chore: rename incorrectly named story
